refactor: 패키지 이동 및 클래스명 수정
