+++ mvc/src/main/java/nextstep/mvc/HandlerAdapterRegistry.java
@@ -1,8 +1,7 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc;
 
 import java.util.ArrayList;
 import java.util.List;
-import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.exception.NotFoundHandlerAdapterException;
 
 public class HandlerAdapterRegistry {
