+++ mvc/src/main/java/nextstep/mvc/HandlerMappingRegistry.java
@@ -1,10 +1,9 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.exception.NotFoundHandlerException;
 
 public class HandlerMappingRegistry {
