+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecutionAdapter.java
@@ -5,7 +5,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.view.ModelAndView;
 
-public class AnnotationHandlerAdapter implements HandlerAdapter {
+public class HandlerExecutionAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
