+++ mvc/src/test/java/nextstep/mvc/HandlerMappingRegistryTest.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -6,6 +6,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import jakarta.servlet.http.HttpServletRequest;
+import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.exception.NotFoundHandlerException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
